python/requirements.txt
py-solc-x
